requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
azure-core==1.29.5
azure-identity==1.15.0
azure-storage-blob==12.19.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
Faker==22.2.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Markdown==3.5.1
msal==1.25.0
msal-extensions==1.0.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
Unidecode==1.3.8
urllib3==2.1.0
zipp==3.17.0
